% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_length.R
\name{sd_length}
\alias{sd_length}
\title{Returns the length of a geometry}
\usage{
sd_length(geom, ...)
}
\arguments{
\item{geom}{(geometry): Input geometry}

\item{...}{For S3 generic compatibility. Must be empty.}
}
\value{
(double)
}
\description{
Returns the length of geom. This function only supports LineString,
MultiLineString, and GeometryCollections containing linear geometries.
Use ST_Perimeter for polygons.
}
\seealso{
\href{https://sedona.apache.org/sedonadb/latest/reference/sql/st_length/}{SedonaDB SQL documentation for ST_Length()}
}
