% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedonafns-package.R
\docType{package}
\name{sedonafns-package}
\alias{sedonafns}
\alias{sedonafns-package}
\title{sedonafns: Bindings for Apache SedonaDB Functions}
\description{
Provides function definitions and documentation for use in 'SedonaDB'.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca}

}
\keyword{internal}
